requirements.txt
# requirements.txt for Enhanced Agentic RAG
google-generativeai>=0.8.0
flask>=2.3.0
flask-cors>=4.0.0
pandas>=1.5.0
numpy>=1.24.0
chromadb>=0.4.0
python-dotenv>=1.0.0
requests>=2.31.0
gspread>=5.12.0
oauth2client>=4.1.3
streamlit>=1.28.0
plotly>=5.18.0 

# For enhanced internet search capabilities
google-search-results>=2.4.2

# Optional: For better text processing
nltk>=3.8
scikit-learn>=1.3.0